unarchive to mem and tinyxml2 xml length fixes from Harri Pitkänen
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk@3659 941e2c2b-deac-454f-805a-451daa25f33c

+++ ospell.cc
@@ -659,7 +659,7 @@ void Speller::build_alphabet_translator(void)
 } // namespace hfst_ol
   
 char*
-hfst_strndup(const char* s, unsigned long n)
+hfst_strndup(const char* s, size_t n)
   {
     char* rv = static_cast<char*>(malloc(sizeof(char)*n+1));
     if (rv == NULL)
